☘️ refactor [#12.3.16]: 2차 개선 - GraphView 렌더링 버그 수정 및 로직 분리(SoC)
- `GraphView.tsx` 캔버스 렌더링 시 fill이 stroke를 덮는 시각적 버그 해결 (실행 순서 보정)
- `GraphContainer.tsx`의 복잡한 상태 관리를 `useGraphData`, `useGraphLiveUpdates` 훅으로 추출하여 단일 책임 원칙(SRP) 적용
- 컨테이너 컴포넌트를 UI 렌더링 위주의 선언적 구조로 개편하여 유지보수성 향상

🔗 Related:
- Issue [#1048]
- @ai-bot-review (https://github.com/jjaayy2222/flownote-mvp/pull/1573#pullrequestreview-4455540956)
